refactor: re-organize Main
