+++ nrs-parser.cabal
@@ -39,6 +39,7 @@ common deps
   other-modules:
     ChapterFile,
     Config,
+    DateUtil,
     FileUtil,
     IndexFile,
     HtmlUtil,
